back_end/requirements.txt
flask
python-dotenv
google-generativeai
flask-cors
PyPDF2
pymongo